test(tree-select): extract reusable test harness
